+++ tools/windows/run_unittests.sh
@@ -100,7 +100,6 @@ diable_wingpu_test="^test_analysis_predictor$|\
 ^test_weight_decay$|\
 ^test_conv2d_int8_mkldnn_op$|\
 ^test_crypto$|\
-^test_callbacks$|\
 ^test_program_prune_backward$|\
 ^test_imperative_ocr_attention_model$|\
 ^test_sentiment$|\
